Policies for everything

See merge request lbaw/lbaw2324/lbaw23155!32

+++ app/Http/Controllers/LikeController.php
@@ -13,6 +13,9 @@ class LikeController extends Controller
 
     public function like($postid)
     {
+        if(!Auth::check() || Auth::user()->cannot('create', Like::class)){
+            return redirect()->back();
+        }
         $like = new Like();
         $like->postid = $postid;
         $like->userid = Auth::id();
@@ -23,6 +26,9 @@ class LikeController extends Controller
     public function dislike($postid)
     {
         $like = Like::where('postid',$postid)->first();
+        if(!Auth::check() || Auth::user()->cannot('delete', $like)){
+            return redirect()->back();
+        }
         $like->delete();
         return redirect()->back();
